requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
pytz==2023.3
python-dotenv==1.0.0
Flask==3.0.0
